Feature/customers

+++ database/migrations/2019_11_22_171834_create_column_city_id_in_sellers.php
@@ -6,10 +6,15 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateColumnCityIdInSellers extends Migration
 {
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
     public function up()
     {
         Schema::table('sellers', function (Blueprint $table) {
-            $table->unsignedBigInteger('city_id')->after('phone');
+            $table->unsignedInteger('city_id')->after('phone');
             $table->foreign('city_id')
                 ->references('id')->on('cities')
                 ->onDelete('cascade')
@@ -17,6 +22,11 @@ class CreateColumnCityIdInSellers extends Migration
         });
     }
 
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
     public function down()
     {
         Schema::table('sellers', function (Blueprint $table) {
